adjusted code to new product ID, commented printing Producers on getAllProducers. Changed monitored values from long to Double

+++ src/main/java/org/example/masterprotocolworker/model/ProducerProposal.java
@@ -9,6 +9,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProducerProposal {
-    private long producerCalculatedRating;
+    private Double producerCalculatedRating;
     private List<Product> products;
 }
